requirements.txt
Django==3.2.4
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.12.0
django-filter==22.1



